Adds `Ord` instance for `HeaderName`.
`HeaderName`s don't really have a sensible ordering so to speak. You
can convert them to their string-equivalents and then compare them,
but at that point, you'd be comparing strings, not `HeaderName`s.
Nevertheless, `Map`s require that keys be an instance of `Ord`, and so
I added an `Ord` instance such that you can now use `HeaderName`s as a
key in a `Map`.