Reverts 68662cfc32c5bafaa3589e5decc59373daa3843a.
While the parsing of individual header values can be left to the user,
it is still convenient to offer the parser a method to more easily
combine repeat headers. Furthermore, the Set-Cookie header is an
exception to the "headers are always strings" rule, and can only be
sensibly handled as a List anyway.